Lift domain sockets tests to a base class
