+++ java/client/test/org/openqa/selenium/remote/http/netty/BUILD.bazel
@@ -8,19 +8,6 @@ java_test_suite(
     deps = [
         "//java/client/src/org/openqa/selenium/remote/http",
         "//java/client/src/org/openqa/selenium/remote/http/netty",
-        "//java/client/test/org/openqa/selenium/testing:test-base",
-        artifact("com.google.guava:guava"),
-        artifact("junit:junit"),
-        artifact("org.assertj:assertj-core"),
-        artifact("io.netty:netty-buffer"),
-        artifact("io.netty:netty-codec-http"),
-        artifact("io.netty:netty-handler"),
-        artifact("io.netty:netty-transport"),
-        artifact("io.netty:netty-transport-native-epoll"),
-        artifact("io.netty:netty-transport-native-epoll-linux-x86_64"),
-        artifact("io.netty:netty-transport-native-kqueue"),
-        artifact("io.netty:netty-transport-native-kqueue-osx-x86_64"),
-        artifact("io.netty:netty-transport-native-unix-common"),
-        artifact("org.slf4j:slf4j-jdk14"),
+        "//java/client/test/org/openqa/selenium/remote/internal:test-lib",
     ],
 )
