+++ java/client/test/org/openqa/selenium/remote/http/netty/NettyClientTest.java
@@ -17,134 +17,13 @@
 
 package org.openqa.selenium.remote.http.netty;
 
-import com.google.common.net.MediaType;
-import io.netty.bootstrap.ServerBootstrap;
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.PooledByteBufAllocator;
-import io.netty.buffer.Unpooled;
-import io.netty.channel.ChannelFuture;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.ChannelInitializer;
-import io.netty.channel.ChannelOption;
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.SimpleChannelInboundHandler;
-import io.netty.channel.epoll.Epoll;
-import io.netty.channel.epoll.EpollEventLoopGroup;
-import io.netty.channel.epoll.EpollServerDomainSocketChannel;
-import io.netty.channel.kqueue.KQueue;
-import io.netty.channel.kqueue.KQueueEventLoopGroup;
-import io.netty.channel.kqueue.KQueueServerDomainSocketChannel;
-import io.netty.channel.unix.DomainSocketAddress;
-import io.netty.channel.unix.DomainSocketChannel;
-import io.netty.channel.unix.ServerDomainSocketChannel;
-import io.netty.handler.codec.http.DefaultFullHttpResponse;
-import io.netty.handler.codec.http.FullHttpRequest;
-import io.netty.handler.codec.http.FullHttpResponse;
-import io.netty.handler.codec.http.HttpObjectAggregator;
-import io.netty.handler.codec.http.HttpResponseStatus;
-import io.netty.handler.codec.http.HttpServerCodec;
-import io.netty.handler.codec.http.HttpServerKeepAliveHandler;
-import io.netty.handler.codec.http.HttpVersion;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.remote.http.ClientConfig;
-import org.openqa.selenium.remote.http.Contents;
 import org.openqa.selenium.remote.http.HttpClient;
-import org.openqa.selenium.remote.http.HttpRequest;
-import org.openqa.selenium.remote.http.HttpResponse;
-import org.openqa.selenium.testing.Safely;
+import org.openqa.selenium.remote.internal.DomainSocketsTestBase;
 
-import java.io.IOException;
-import java.net.SocketAddress;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.concurrent.atomic.AtomicReference;
+public class NettyClientTest extends DomainSocketsTestBase {
 
-import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_LENGTH;
-import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_TYPE;
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assumptions.assumeThat;
-import static org.openqa.selenium.remote.http.HttpMethod.GET;
-
-public class NettyClientTest {
-
-  private EventLoopGroup group;
-  private ChannelFuture future;
-  private URI socket;
-  private AtomicReference<String> responseText = new AtomicReference<>();
-
-  @Before
-  public void setupUnixDomainSocketServer() throws IOException, URISyntaxException {
-    Class<? extends ServerDomainSocketChannel> channelType = null;
-
-    if (Epoll.isAvailable()) {
-      group = new EpollEventLoopGroup(2);
-      channelType = EpollServerDomainSocketChannel.class;
-    } else if (KQueue.isAvailable()) {
-      group = new KQueueEventLoopGroup(2);
-      channelType = KQueueServerDomainSocketChannel.class;
-    }
-
-    assumeThat(group).isNotNull();
-    assumeThat(channelType).isNotNull();
-
-    ServerBootstrap bootstrap = new ServerBootstrap()
-      .group(group)
-      .option(ChannelOption.SO_BACKLOG, 1024)
-      .childOption(ChannelOption.ALLOCATOR, PooledByteBufAllocator.DEFAULT)
-      .channel(channelType)
-      .childHandler(new ChannelInitializer<DomainSocketChannel>() {
-        @Override
-        protected void initChannel(DomainSocketChannel ch) {
-          ch.pipeline()
-            .addLast("http-codec", new HttpServerCodec())
-            .addLast("http-keep-alive", new HttpServerKeepAliveHandler())
-            .addLast("http-aggregator", new HttpObjectAggregator(Integer.MAX_VALUE))
-            .addLast(new SimpleChannelInboundHandler<FullHttpRequest>() {
-              @Override
-              protected void channelRead0(ChannelHandlerContext ctx, FullHttpRequest req) {
-                byte[] bytes = responseText.get().getBytes(UTF_8);
-                ByteBuf text = Unpooled.wrappedBuffer(bytes);
-                FullHttpResponse res = new DefaultFullHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK, text);
-                res.headers().set(CONTENT_TYPE, MediaType.PLAIN_TEXT_UTF_8.toString());
-                res.headers().set(CONTENT_LENGTH, bytes.length);
-
-                ctx.writeAndFlush(res);
-              }
-            });
-        }
-      });
-
-    Path temp = Files.createTempFile("domain-socket-test", "socket");
-    Files.deleteIfExists(temp);
-
-    SocketAddress address = new DomainSocketAddress(temp.toFile());
-    future = bootstrap.bind(address);
-
-    socket = new URI("unix", null, null, 0, temp.toString(), null, null);
+  @Override
+  protected HttpClient.Factory createFactory() {
+    return new NettyClient.Factory();
   }
-
-  @After
-  public void shutdown() {
-    Safely.safelyCall(() -> group.shutdownGracefully());
-    Safely.safelyCall(() -> future.channel().closeFuture().sync());
-  }
-
-  @Test
-  public void shouldBeAbleToConnectToAUnixDomainSocketUrl() {
-    ClientConfig config = ClientConfig.defaultConfig().baseUri(socket);
-    HttpClient client = new NettyClient.Factory().createClient(config);
-
-    String emphaticCheeseEnjoyment = "I like cheese!";
-    responseText.set(emphaticCheeseEnjoyment);
-
-    HttpResponse res = client.execute(new HttpRequest(GET, "/do-you-like-cheese"));
-
-    assertThat(Contents.string(res)).isEqualTo(emphaticCheeseEnjoyment);
-  }
-
 }
